refactor payment page
